--- a/src/cryptonote_basic/account.cpp
+++ b/src/cryptonote_basic/account.cpp
@@ -131,7 +131,7 @@ DISABLE_VS_WARNINGS(4244 4345)
   std::string account_base::get_public_address_str(bool testnet) const
   {
     //TODO: change this code into base 58
-    return get_account_address_as_str(testnet, m_keys.m_account_address);
+    return get_account_address_as_str(testnet, false, m_keys.m_account_address);
   }
   //-----------------------------------------------------------------
   std::string account_base::get_public_integrated_address_str(const crypto::hash8 &payment_id, bool testnet) const