@JollyMort 2017-05-30
Remember, funds never reside on some address. Fundamentally, all the outputs belong to you because you can recover the individual private keys and they're not associated with an address. If the sender lost the relevant TX key, nobody can ever prove an output was sent to some address unless you give them the view key. Even if you **know** from your records you sent x monero to address y in TX z, you can't ever prove it without TX key or view key. Grouping them into addresses or sub-addresses is kind of arbitrary. But yeah, it could be useful to draw borders and avoid combining some together, and have the sweep_all to collect from all subs (does it do that already?)