@xmrgit 2017-05-30
just registered to ask if we can have a `sweep_all` for every sub-address instead of just `move_sub2main`

so more something like `sweep_all_sub 2 0` which will move all funds from sub-address 2 to sub-address 0.

I like to sweep my current sub-wallets usually more than 1 time before I send them to my main-wallets for privacy reasons, that's why I asked.