@Jaqueeee 2017-05-31
Good work! What's the impact on performance when rescanning blockchain? Has it been benchmarked?