@dan-da 2017-06-01
> Error: No outputs found, or daemon is not ready

Can this error message be improved?   Even as  a technical user, this would mislead me.

Perhaps something like:  "Error: No single sub-address contains adequate funds although wallet balance is sufficient.   Suggestion: aggregate funds to a fresh address."