@kenshi84 2017-06-03
Thank you guys for your feedback!

@xmrgit @knaccc 
I agree with your point, so I dropped the command `move_sub2main` now and instead modified the command `sweep_all` such that it optionally accepts the index of subaddress to be swept. Please check the updated OP above, I hope it achieves what you had in mind.

@NanoAkron 

> Is there a possibility of sweeping all subaddress funds back into my main address? 

Yes, it's possible to sweep all the balance of each subaddress to the main address _individually_. Alternatively, although not recommended for privacy reasons, it's also possible to sweep all the balances of all the subaddresses to the main address in a single transaction if you switch off the safeguard by doing `set allow-combine-subaddresses 1`.

> Is that how bitcoin's HD wallet works? 

Conceptually it's kind of similar to Bitcoin's HD wallet, but the actual problem being addressed here is vastly different because in Monero, wallet addresses never appear in the blockchain as @JollyMort stressed.

> Or would that be more like the 'disposable address' scheme...?

I couldn't understand your question...


@JollyMort 

> unless he was the one who sent to B

This is the concern addressed by the "not combining subaddresses" scheme. If Alice sent some funds to two different subaddresses and saw a transaction later that included those two outputs in the input ring signatures, she would guess that the two subaddresses belong to the same person.

> it could be useful to draw borders and avoid combining some together, and have the sweep_all to collect from all subs (does it do that already?)

It's not recommended (but still possible) to sweep all the funds in all the subaddresses to the main address at once, due to the above reason.


@Jaqueeee

> What's the impact on performance when rescanning blockchain? Has it been benchmarked?

No, it's not benchmarked yet. But my guess is that the additional cost for blockchain scanning will be really small (i.e. 1 hashtable lookup, 1 hash computation, and 1 scalar addition, for each incoming tranfer).

@dan-da

Thanks a lot for your suggestion! I changed the error message following your suggestion, please see the updated OP.