@stoffu 2017-06-03
> Yup. But I could by chance make a ring signature using some 2 inputs to which Alice sent funds (to someone else, not me) in my ring and she could come to the same (but wrong) conclusion :) Plausible deniability FTW

The point is that the likelihood of such a situation is fairly small, and if that happens again and again, the plausible deniability will diminish pretty quickly.