@xmrgit 2017-06-04
would it also make sense to modify the commant `transfer` such that it optionally accepts the index of a subaddress instead of only randomly choosing a subaddress to be spent?