@kenshi84 2017-06-08
@xmrgit @knaccc 
The CLI has been updated quite a bit, please check the edited OP (after section 3).