@knaccc 2017-06-08
Minor note: I think it might help if the command syntax:
`address [ <index_min> [<index_max>]|fresh ]`
was instead written as:
`address [ [<index_min> [<index_max>]] | fresh ]`