@xmrgit 2017-06-08
@kenshi84 
I love your work. but one question: does it make sense to have the `<index=n>` in `sweep_all` command as optional instead of required?